refactor: Post 비즈니스프로필과 일반 유저가 하는 역할에 따라 파일 분리(#162)
