+++ backend/src/main/java/com/minibeit/post/domain/repository/PostApplicantRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.minibeit.post.domain.repository;
 
+import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.post.domain.PostApplicant;
 import com.minibeit.post.dto.PostApplicantDto;
 
@@ -9,9 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface PostApplicantRepositoryCustom {
-    List<PostApplicantDto.UserInfo> findAllByPostAndDoDate(Long postId, LocalDate doDate);
-
-    List<PostApplicantDto.UserInfo> findAllByPostAndDoDateAndApprove(Long postId, LocalDate doDate);
+    List<PostApplicantDto.UserInfo> findAllByPostAndDoDate(Long postId, ApplyStatus applyStatus, LocalDate doDate);
 
     List<PostApplicant> findAllByApplyStatusIsWait(Long postId);
