+++ backend/src/main/java/com/minibeit/post/domain/repository/PostApplicantRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.post.domain.PostApplicant;
 import com.minibeit.post.dto.PostApplicantDto;
 import com.minibeit.post.dto.QPostApplicantDto_UserInfo;
+import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 
@@ -22,7 +23,7 @@ public class PostApplicantRepositoryImpl implements PostApplicantRepositoryCusto
     private final JPAQueryFactory queryFactory;
 
     @Override
-    public List<PostApplicantDto.UserInfo> findAllByPostAndDoDate(Long postId, LocalDate doDate) {
+    public List<PostApplicantDto.UserInfo> findAllByPostAndDoDate(Long postId, ApplyStatus applyStatus, LocalDate doDate) {
         return queryFactory.select(new QPostApplicantDto_UserInfo(
                         user.id, user.name, user.birth, user.gender, user.phoneNum, user.job, post.doTime, postApplicant.applyStatus, postApplicant.businessFinish, postDoDate.id, postDoDate.doDate
                 ))
@@ -34,27 +35,19 @@ public class PostApplicantRepositoryImpl implements PostApplicantRepositoryCusto
                         .and(postDoDate.doDate.year().eq(doDate.getYear())
                                 .and(postDoDate.doDate.month().eq(doDate.getMonthValue()))
                                 .and(postDoDate.doDate.dayOfMonth().eq(doDate.getDayOfMonth())))
-                        .and(postApplicant.applyStatus.ne(ApplyStatus.REJECT)))
+                        .and(byApplyStatus(applyStatus)))
                 .orderBy(postDoDate.doDate.asc())
                 .fetch();
     }
 
-    @Override
-    public List<PostApplicantDto.UserInfo> findAllByPostAndDoDateAndApprove(Long postId, LocalDate doDate) {
-        return queryFactory.select(new QPostApplicantDto_UserInfo(
-                        user.id, user.name, user.birth, user.gender, user.phoneNum, user.job, post.doTime, postApplicant.applyStatus, postApplicant.businessFinish, postDoDate.id, postDoDate.doDate
-                ))
-                .from(postApplicant)
-                .join(postApplicant.user, user)
-                .join(postApplicant.postDoDate, postDoDate)
-                .join(postDoDate.post, post)
-                .where(post.id.eq(postId)
-                        .and(postDoDate.doDate.year().eq(doDate.getYear())
-                                .and(postDoDate.doDate.month().eq(doDate.getMonthValue()))
-                                .and(postDoDate.doDate.dayOfMonth().eq(doDate.getDayOfMonth())))
-                        .and(postApplicant.applyStatus.eq(ApplyStatus.APPROVE)))
-                .orderBy(postDoDate.doDate.asc())
-                .fetch();
+    private BooleanExpression byApplyStatus(ApplyStatus applyStatus) {
+        if (applyStatus.equals(ApplyStatus.WAIT)) {
+            return postApplicant.applyStatus.ne(ApplyStatus.REJECT);
+        }
+        if (applyStatus.equals(ApplyStatus.APPROVE)) {
+            return postApplicant.applyStatus.eq(ApplyStatus.APPROVE);
+        }
+        return null;
     }
 
     @Override
