+++ backend/src/main/java/com/minibeit/post/service/PostApplicantService.java
@@ -1,22 +1,19 @@
 package com.minibeit.post.service;
 
-import com.minibeit.businessprofile.domain.repository.UserBusinessProfileRepository;
 import com.minibeit.common.exception.PermissionException;
-import com.minibeit.post.domain.*;
+import com.minibeit.post.domain.ApplyStatus;
+import com.minibeit.post.domain.Post;
+import com.minibeit.post.domain.PostApplicant;
+import com.minibeit.post.domain.PostDoDate;
 import com.minibeit.post.domain.repository.PostApplicantRepository;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
-import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.PostApplicantDto;
-import com.minibeit.post.dto.PostApplicantRequest;
-import com.minibeit.post.dto.PostApplicantResponse;
 import com.minibeit.post.service.exception.*;
 import com.minibeit.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -27,8 +24,6 @@ public class PostApplicantService {
     private final PostRepository postRepository;
     private final PostDoDateRepository postDoDateRepository;
     private final PostApplicantRepository postApplicantRepository;
-    private final UserBusinessProfileRepository userBusinessProfileRepository;
-    private final RejectPostRepository rejectPostRepository;
 
     public void apply(Long postId, Long postDoDateId, User user) {
         PostDoDate postDoDate = postDoDateRepository.findById(postDoDateId).orElseThrow(PostDoDateNotFoundException::new);
@@ -36,7 +31,10 @@ public class PostApplicantService {
             throw new DuplicateApplyException();
         }
         Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-        postDoDateFullCheck(post, postDoDate);
+        List<PostApplicant> postApplicants = postDoDate.getPostApplicantList();
+        if (!post.applyPossible(postApplicants)) {
+            throw new PostIsFullException();
+        }
 
         PostApplicant postApplicant = PostApplicant.create(postDoDate, user);
         postApplicantRepository.save(postApplicant);
@@ -52,38 +50,6 @@ public class PostApplicantService {
         postApplicant.updateMyFinish();
     }
 
-    public void applyApprove(Long postId, Long postDoDateId, Long userId, User user) {
-        Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-        permissionCheck(user, post);
-        PostDoDate postDoDate = postDoDateRepository.findById(postDoDateId).orElseThrow(PostDoDateNotFoundException::new);
-        postDoDateFullCheck(post, postDoDate);
-
-        PostApplicant postApplicant = postApplicantRepository.findByPostDoDateIdAndUserId(postDoDateId, userId).orElseThrow(PostApplicantNotFoundException::new);
-
-        postApplicant.updateStatus(ApplyStatus.APPROVE);
-
-        List<PostApplicant> approvedPostApplicant = postApplicantRepository.findAllByPostDoDateIdAndStatusIsApprove(postDoDateId);
-        postDoDate.updateFull(approvedPostApplicant);
-    }
-
-    public void applyApproveCancel(Long postId, Long postDoDateId, Long userId, User user) {
-        Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-        permissionCheck(user, post);
-        PostApplicant postApplicant = postApplicantRepository.findByPostDoDateIdAndUserId(postDoDateId, userId).orElseThrow(PostApplicantNotFoundException::new);
-        postApplicant.updateStatus(ApplyStatus.WAIT);
-    }
-
-    public void applyReject(Long postId, Long postDoDateId, Long userId, PostApplicantRequest.ApplyReject request, User user) {
-        Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-        permissionCheck(user, post);
-        PostApplicant postApplicant = postApplicantRepository.findByPostDoDateIdAndUserIdWithPostDoDate(postDoDateId, userId).orElseThrow(PostApplicantNotFoundException::new);
-
-        postApplicant.updateStatus(ApplyStatus.REJECT);
-
-        RejectPost rejectPost = RejectPost.create(post.getTitle(), post.getPlace(), post.getContact(), post.getDoTime(), postApplicant.getPostDoDate().getDoDate(), request.getComment(), postApplicant.getUser());
-        rejectPostRepository.save(rejectPost);
-    }
-
     public void applyCancel(Long postDoDateId, User user) {
         postApplicantRepository.deleteByPostDoDateIdAndUserId(postDoDateId, user.getId());
 
@@ -91,38 +57,5 @@ public class PostApplicantService {
 
         List<PostApplicant> approvedPostApplicant = postApplicantRepository.findAllByPostDoDateIdAndStatusIsApprove(postDoDateId);
         postDoDate.updateFull(approvedPostApplicant);
-    }
-
-    public void attendChange(Long postId, Long postDoDateId, Long userId, PostApplicantRequest.AttendChange request, User user) {
-        Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
-        permissionCheck(user, post);
-        PostApplicant postApplicant = postApplicantRepository.findByPostDoDateIdAndUserId(postDoDateId, userId).orElseThrow(PostApplicantNotFoundException::new);
-
-        postApplicant.changeBusinessFinish(request.getIsAttend());
-    }
-
-    @Transactional(readOnly = true)
-    public List<PostApplicantResponse.ApplicantInfo> getApplicantListByDate(Long postId, LocalDate doDate) {
-        List<PostApplicantDto.UserInfo> userInfoList = postApplicantRepository.findAllByPostAndDoDate(postId, doDate);
-        return PostApplicantResponse.ApplicantInfo.dtoToResponse(userInfoList);
-    }
-
-    @Transactional(readOnly = true)
-    public List<PostApplicantResponse.ApplicantInfo> getApproveApplicantListByDate(Long postId, LocalDate doDate) {
-        List<PostApplicantDto.UserInfo> userInfoList = postApplicantRepository.findAllByPostAndDoDateAndApprove(postId, doDate);
-        return PostApplicantResponse.ApplicantInfo.dtoToResponse(userInfoList);
-    }
-
-    private void permissionCheck(User user, Post post) {
-        if (!userBusinessProfileRepository.existsByUserIdAndBusinessProfileId(user.getId(), post.getBusinessProfile().getId())) {
-            throw new PermissionException();
-        }
-    }
-
-    private void postDoDateFullCheck(Post post, PostDoDate postDoDate) {
-        List<PostApplicant> postApplicants = postApplicantRepository.findAllByPostDoDateIdAndStatusIsApprove(postDoDate.getId());
-        if (!post.applyPossible(postApplicants)) {
-            throw new PostIsFullException();
-        }
     }
 }
