+++ backend/src/main/java/com/minibeit/post/web/PostApplicantController.java
@@ -1,17 +1,14 @@
 package com.minibeit.post.web;
 
-import com.minibeit.post.dto.PostApplicantRequest;
-import com.minibeit.post.dto.PostApplicantResponse;
 import com.minibeit.post.service.PostApplicantService;
 import com.minibeit.security.userdetails.CurrentUser;
 import com.minibeit.security.userdetails.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.time.LocalDate;
-import java.util.List;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
@@ -35,48 +32,5 @@ public class PostApplicantController {
     public ResponseEntity<Void> applyCancel(@PathVariable Long postDoDateId, @CurrentUser CustomUserDetails customUserDetails) {
         postApplicantService.applyCancel(postDoDateId, customUserDetails.getUser());
         return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{postId}/date/{postDoDateId}/apply/approve/{userId}")
-    public ResponseEntity<Void> applyApprove(@PathVariable Long postId, @PathVariable Long postDoDateId,
-                                             @PathVariable Long userId, @CurrentUser CustomUserDetails customUserDetails) {
-        postApplicantService.applyApprove(postId, postDoDateId, userId, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{postId}/date/{postDoDateId}/apply/approve/cancel/{userId}")
-    public ResponseEntity<Void> applyApproveCancel(@PathVariable Long postId, @PathVariable Long postDoDateId,
-                                                   @PathVariable Long userId, @CurrentUser CustomUserDetails customUserDetails) {
-        postApplicantService.applyApproveCancel(postId, postDoDateId, userId, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{postId}/date/{postDoDateId}/apply/reject/{userId}")
-    public ResponseEntity<Void> applyReject(@PathVariable Long postId, @PathVariable Long postDoDateId,
-                                            @PathVariable Long userId, @RequestBody PostApplicantRequest.ApplyReject request,
-                                            @CurrentUser CustomUserDetails customUserDetails) {
-        postApplicantService.applyReject(postId, postDoDateId, userId, request, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{postId}/date/{postDoDateId}/attend/change/{userId}")
-    public ResponseEntity<Void> attendChange(@PathVariable Long postId, @PathVariable Long postDoDateId,
-                                             @PathVariable Long userId, @RequestBody PostApplicantRequest.AttendChange request, @CurrentUser CustomUserDetails customUserDetails) {
-        postApplicantService.attendChange(postId, postDoDateId, userId, request, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/{postId}/applicant/list")
-    public ResponseEntity<List<PostApplicantResponse.ApplicantInfo>> applicantListByDate(@PathVariable Long postId,
-                                                                                         @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate doDate) {
-        List<PostApplicantResponse.ApplicantInfo> response = postApplicantService.getApplicantListByDate(postId, doDate);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @GetMapping("/{postId}/applicant/confirm/list")
-    public ResponseEntity<List<PostApplicantResponse.ApplicantInfo>> approveApplicantListByDate(@PathVariable Long postId,
-                                                                                                @RequestParam @DateTimeFormat(pattern = "yyyy-MM-dd") LocalDate doDate) {
-        List<PostApplicantResponse.ApplicantInfo> response = postApplicantService.getApproveApplicantListByDate(postId, doDate);
-        return ResponseEntity.ok().body(response);
     }
 }
