+++ backend/src/main/java/com/minibeit/post/web/PostController.java
@@ -3,24 +3,18 @@ package com.minibeit.post.web;
 import com.minibeit.common.dto.PageDto;
 import com.minibeit.post.domain.ApplyStatus;
 import com.minibeit.post.domain.Payment;
-import com.minibeit.post.domain.PostStatus;
-import com.minibeit.post.dto.PostRequest;
 import com.minibeit.post.dto.PostResponse;
-import com.minibeit.post.service.PostFileService;
 import com.minibeit.post.service.PostService;
 import com.minibeit.security.userdetails.CurrentUser;
 import com.minibeit.security.userdetails.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.net.URI;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.YearMonth;
 import java.util.List;
 
 @RestController
@@ -28,31 +22,10 @@ import java.util.List;
 @RequestMapping("/api/post")
 public class PostController {
     private final PostService postService;
-    private final PostFileService postFileService;
-
-    @PostMapping("/info")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<PostResponse.OnlyId> createInfo(@RequestBody PostRequest.CreateInfo request, @CurrentUser CustomUserDetails customUserDetails) {
-        PostResponse.OnlyId response = postService.createInfo(request, customUserDetails.getUser());
-        return ResponseEntity.created(URI.create("/api/post/" + response.getId())).body(response);
-    }
-
-    @PostMapping("/{postId}/files")
-    @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<PostResponse.OnlyId> addFiles(@PathVariable Long postId, PostRequest.AddFile request, @CurrentUser CustomUserDetails customUserDetails) {
-        PostResponse.OnlyId response = postFileService.addFiles(postId, request, customUserDetails.getUser());
-        return ResponseEntity.created(URI.create("/api/post/" + response.getId())).body(response);
-    }
 
     @PostMapping("/{postId}/like")
     public ResponseEntity<Void> like(@PathVariable Long postId, @CurrentUser CustomUserDetails customUserDetails) {
         postService.createOrDeletePostLike(postId, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/{postId}/completed")
-    public ResponseEntity<Void> recruitmentCompleted(@PathVariable Long postId, @CurrentUser CustomUserDetails customUserDetails) {
-        postService.recruitmentCompleted(postId, customUserDetails.getUser());
         return ResponseEntity.ok().build();
     }
 
@@ -89,12 +62,6 @@ public class PostController {
         return ResponseEntity.ok().body(response);
     }
 
-    @GetMapping("/myComplete/list")
-    public ResponseEntity<Page<PostResponse.GetMyCompletedList>> getListByMyCompleteList(PageDto pageDto, @CurrentUser CustomUserDetails customUserDetails) {
-        Page<PostResponse.GetMyCompletedList> response = postService.getListByMyCompleteList(customUserDetails.getUser(), pageDto);
-        return ResponseEntity.ok().body(response);
-    }
-
     @GetMapping("/apply/list")
     public ResponseEntity<Page<PostResponse.GetMyApplyList>> getListByApplyStatus(@RequestParam(name = "status") ApplyStatus applyStatus,
                                                                                   PageDto pageDto,
@@ -103,31 +70,9 @@ public class PostController {
         return ResponseEntity.ok().body(response);
     }
 
-    @GetMapping("/business/profile/{businessProfileId}/list")
-    public ResponseEntity<Page<PostResponse.GetListByBusinessProfile>> getListByBusinessProfile(@PathVariable Long businessProfileId,
-                                                                                                @RequestParam(defaultValue = "RECRUIT", name = "status") PostStatus postStatus,
-                                                                                                PageDto pageDto) {
-        Page<PostResponse.GetListByBusinessProfile> response = postService.getListByBusinessProfile(businessProfileId, postStatus, pageDto);
+    @GetMapping("/myComplete/list")
+    public ResponseEntity<Page<PostResponse.GetMyCompletedList>> getListByMyCompleteList(PageDto pageDto, @CurrentUser CustomUserDetails customUserDetails) {
+        Page<PostResponse.GetMyCompletedList> response = postService.getListByMyCompleteList(customUserDetails.getUser(), pageDto);
         return ResponseEntity.ok().body(response);
-    }
-
-    @GetMapping("/{postId}/exist/doDate/list")
-    public ResponseEntity<PostResponse.DoDateList> getDoDateList(@PathVariable Long postId,
-                                                                 @RequestParam(name = "yearMonth") @DateTimeFormat(pattern = "yyyy-MM") YearMonth yearMonth) {
-        PostResponse.DoDateList response = postService.getDoDateListByYearMonth(postId, yearMonth);
-        return ResponseEntity.ok().body(response);
-    }
-
-    @PutMapping("/{postId}")
-    public ResponseEntity<PostResponse.OnlyId> updateContent(@PathVariable Long postId, @RequestBody PostRequest.UpdateContent request,
-                                                             @CurrentUser CustomUserDetails customUserDetails) {
-        PostResponse.OnlyId response = postService.updateContent(postId, request, customUserDetails.getUser());
-        return ResponseEntity.ok().body(response);
-    }
-
-    @DeleteMapping("/{postId}")
-    public ResponseEntity<Void> deleteOne(@PathVariable Long postId, @CurrentUser CustomUserDetails customUserDetails) {
-        postService.deleteOne(postId, customUserDetails.getUser());
-        return ResponseEntity.ok().build();
     }
 }
