+++ backend/src/test/java/com/minibeit/post/web/PostControllerTest.java
@@ -4,10 +4,7 @@ import com.minibeit.MvcTest;
 import com.minibeit.avatar.domain.Avatar;
 import com.minibeit.businessprofile.domain.BusinessProfile;
 import com.minibeit.post.domain.*;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
 import com.minibeit.post.dto.PostResponse;
-import com.minibeit.post.service.PostFileService;
 import com.minibeit.post.service.PostService;
 import com.minibeit.school.domain.School;
 import com.minibeit.security.userdetails.CustomUserDetails;
@@ -17,18 +14,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.core.io.ClassPathResource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.MediaType;
-import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.ResultActions;
 
-import java.io.InputStream;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.*;
 
@@ -46,8 +39,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 class PostControllerTest extends MvcTest {
     @MockBean
     private PostService postService;
-    @MockBean
-    private PostFileService postFileService;
 
     private BusinessProfile businessProfile;
     private Post post1;
@@ -111,131 +102,6 @@ class PostControllerTest extends MvcTest {
         postList.add(post2);
         postDoDate1 = PostDoDate.builder().id(1L).full(false).post(post1).doDate(LocalDateTime.of(2021, 9, 5, 9, 30)).build();
         postDoDate2 = PostDoDate.builder().id(2L).full(true).post(post1).doDate(LocalDateTime.of(2021, 9, 5, 10, 30)).build();
-    }
-
-    @Test
-    @DisplayName("게시물 정보입력(생성) 문서화")
-    public void createInfo() throws Exception {
-        PostRequest.CreateInfo request = PostRequest.CreateInfo.builder()
-                .title("커피를 얼마나 마셔야 잠을 못잘까~?")
-                .content("실험 내용")
-                .place("고려대학교 연구실")
-                .contact("010-1234-1234")
-                .category("미디어")
-                .headcount(10)
-                .payment(Payment.CACHE)
-                .cache(10000)
-                .goods(null)
-                .paymentDetail("계좌로 지급해드립니다.")
-                .condition(true)
-                .conditionDetail("커피 많이 드시는 사람|")
-                .doTime(60)
-                .schoolId(1L)
-                .businessProfileId(1L)
-                .startDate(LocalDateTime.of(2021, 9, 26, 17, 30))
-                .endDate(LocalDateTime.of(2021, 10, 2, 17, 30))
-                .doDateList(Collections.singletonList(PostDto.PostDoDate.builder().groupId(1).doDate(LocalDateTime.of(2021, 9, 26, 17, 30)).build()))
-                .build();
-        PostResponse.OnlyId response = PostResponse.OnlyId.builder().id(1L).build();
-
-        given(postService.createInfo(any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(post("/api/post/info")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request))
-                .characterEncoding("UTF-8")
-        );
-
-        results.andExpect(status().isCreated())
-                .andDo(print())
-                .andDo(document("post-create-info",
-                        requestFields(
-                                fieldWithPath("title").type(JsonFieldType.STRING).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("content").type(JsonFieldType.STRING).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("place").type(JsonFieldType.STRING).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("contact").type(JsonFieldType.STRING).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("category").type(JsonFieldType.STRING).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("headcount").type(JsonFieldType.NUMBER).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("payment").type(JsonFieldType.STRING).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("cache").description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("goods").description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("paymentDetail").type(JsonFieldType.STRING).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("condition").type(JsonFieldType.BOOLEAN).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("conditionDetail").description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("doTime").type(JsonFieldType.NUMBER).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("schoolId").type(JsonFieldType.NUMBER).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("businessProfileId").type(JsonFieldType.NUMBER).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("startDate").type(JsonFieldType.STRING).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("endDate").type(JsonFieldType.STRING).description("모집 시작 날짜 및 시간"),
-                                fieldWithPath("doDateList[].groupId").type(JsonFieldType.NUMBER).description("모집 마감 날짜 및 시간"),
-                                fieldWithPath("doDateList[].doDate").type(JsonFieldType.STRING).description("참여 가능 날짜(시간포함)")
-                        ),
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("생성된 게시물 식별자")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("게시물 파일 추가")
-    public void addFiles() throws Exception {
-        InputStream is = new ClassPathResource("mock/images/enjoy.png").getInputStream();
-        MockMultipartFile files = new MockMultipartFile("files", "avatar.jpg", "image/jpg", is.readAllBytes());
-        PostResponse.OnlyId response = PostResponse.OnlyId.builder().id(1L).build();
-
-        given(postFileService.addFiles(any(), any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(
-                fileUpload("/api/post/{postId}/files", 1)
-                        .file(files)
-                        .contentType(MediaType.MULTIPART_FORM_DATA)
-                        .characterEncoding("UTF-8")
-        );
-
-        results.andExpect(status().isCreated())
-                .andDo(document("post-add-files",
-                        pathParameters(
-                                parameterWithName("postId").description("게시물 식별자")
-                        ),
-                        requestParts(
-                                partWithName("files").description("게시물에 추가할 파일")
-                        ),
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("게시물 식별자")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("게시물 세부내용 수정")
-    public void updateContent() throws Exception {
-        PostRequest.UpdateContent request = PostRequest.UpdateContent.builder()
-                .updatedContent("수정된 내용 추가")
-                .build();
-        PostResponse.OnlyId response = PostResponse.OnlyId.builder().id(1L).build();
-
-        given(postService.updateContent(any(), any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders
-                .put("/api/post/{postId}", 1)
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request))
-                .characterEncoding("UTF-8")
-        );
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-update-content",
-                        pathParameters(
-                                parameterWithName("postId").description("게시물 식별자")
-                        ),
-                        requestFields(
-                                fieldWithPath("updatedContent").type(JsonFieldType.STRING).description("수정된 내용")
-                        ),
-                        responseFields(
-                                fieldWithPath("id").type(JsonFieldType.NUMBER).description("게시물 식별자")
-                        )
-                ));
     }
 
     @Test
@@ -533,100 +399,6 @@ class PostControllerTest extends MvcTest {
                                 fieldWithPath("totalElements").description("전체 개수"),
                                 fieldWithPath("last").description("마지막 페이지인지 식별"),
                                 fieldWithPath("totalPages").description("전체 페이지")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("비즈니스 프로필로 생성한 실험 리스트 문서화")
-    public void getListByBusinessProfile() throws Exception {
-        Page<Post> postPage = new PageImpl<>(postList, PageRequest.of(1, 6), postList.size());
-        Page<PostResponse.GetListByBusinessProfile> response = postPage.map(PostResponse.GetListByBusinessProfile::build);
-
-        given(postService.getListByBusinessProfile(any(), any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders
-                .get("/api/post/business/profile/{businessProfileId}/list", 1)
-                .param("page", "1")
-                .param("size", "5")
-                .param("status", "RECRUIT"));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-getList-business-profile",
-                        pathParameters(
-                                parameterWithName("businessProfileId").description("비즈니스 프로필 식별자")
-                        ),
-                        requestParameters(
-                                parameterWithName("page").description("조회할 페이지"),
-                                parameterWithName("size").description("조회할 사이즈"),
-                                parameterWithName("status").description("RECRUIT(모집중) or COMPLETE(모집완료)")
-                        ),
-                        relaxedResponseFields(
-                                fieldWithPath("content[].id").type(JsonFieldType.NUMBER).description("게시물 식별자"),
-                                fieldWithPath("content[].title").type(JsonFieldType.STRING).description("제목"),
-                                fieldWithPath("content[].likes").type(JsonFieldType.NUMBER).description("즐겨찾기 수"),
-                                fieldWithPath("totalElements").description("전체 개수"),
-                                fieldWithPath("last").description("마지막 페이지인지 식별"),
-                                fieldWithPath("totalPages").description("전체 페이지")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("게시물 실험이 있는 날짜 목록 조회(년,월 기준) 문서화")
-    public void getDoDateList() throws Exception {
-        Set<LocalDate> localDates = new HashSet<>();
-        localDates.add(LocalDate.of(2021, 9, 21));
-        localDates.add(LocalDate.of(2021, 9, 22));
-        localDates.add(LocalDate.of(2021, 9, 23));
-        PostResponse.DoDateList response = PostResponse.DoDateList.builder().doDateList(localDates).build();
-
-        given(postService.getDoDateListByYearMonth(any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders
-                .get("/api/post/{postId}/exist/doDate/list", 1)
-                .param("yearMonth", "2021-09"));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-doDate-list",
-                        pathParameters(
-                                parameterWithName("postId").description("게시물 식별자")
-                        ),
-                        requestParameters(
-                                parameterWithName("yearMonth").description("조회할 날짜")
-                        ),
-                        responseFields(
-                                fieldWithPath("doDateList[]").type(JsonFieldType.ARRAY).description("실험 있는 날짜")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("게시물 삭제 문서화")
-    public void deleteOne() throws Exception {
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders.delete("/api/post/{postId}", 1));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-deleteOne",
-                        pathParameters(
-                                parameterWithName("postId").description("삭제할 게시물 식별자")
-                        )
-                ));
-    }
-
-    @Test
-    @DisplayName("게시물 모집상태 변화 문서화")
-    public void completed() throws Exception {
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders.post("/api/post/{postId}/completed", 1));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-recruitment-Completed",
-                        pathParameters(
-                                parameterWithName("postId").description("모집완료 할 게시물 식별자")
                         )
                 ));
     }
